+++ packages/core/src/hooks/index.ts
@@ -1,17 +1,17 @@
-export * from "./useBlock";
-export * from "./useBlockNumber";
 export * from "./useAccount";
 export * from "./useBalance";
+export * from "./useBlock";
+export * from "./useBlockNumber";
 export * from "./useConnect";
 export * from "./useContractFactory";
-export * from "./useContractRead";
 export * from "./useContractWrite";
-export * from "./useExplorer";
 export * from "./useDeployAccount";
 export * from "./useDisconnect";
+export * from "./useExplorer";
 export * from "./useInvalidateOnBlock";
 export * from "./useNetwork";
 export * from "./useProvider";
+export * from "./useReadContract";
 export * from "./useSign";
 // export * from "./useStarkAddress";
 // export * from "./useStarkName";
