+++ packages/core/src/hooks/useReadContract.test-d.ts
@@ -1,8 +1,8 @@
-import { tokenAddress } from "./../../test/devnet";
-import { testAbi } from "../../test/test-abi";
-import { assertType, describe, it } from "vitest";
-import { UseContractReadProps, useContractRead } from "./useContractRead";
 import { FunctionRet } from "abi-wan-kanabi/dist/kanabi";
+import { assertType, describe, it } from "vitest";
+import { tokenAddress } from "../../test/devnet";
+import { testAbi } from "../../test/test-abi";
+import { UseReadContractProps, useReadContract } from "./useReadContract";
 
 type TAbi = typeof testAbi;
 
@@ -11,14 +11,14 @@ const commonProps = {
   address: tokenAddress,
 } as const;
 
-describe("Types test for useContractRead.ts", () => {
+describe("Types test for useReadContract.ts", () => {
   it("correct function name for given abi", () => {
     const props = {
       ...commonProps,
       functionName: "fn_felt",
       args: 1234,
     } as const;
-    assertType<UseContractReadProps<TAbi, "fn_felt">>(props);
+    assertType<UseReadContractProps<TAbi, "fn_felt">>(props);
   });
 
   it("wrong function name for given abi", () => {
@@ -29,7 +29,7 @@ describe("Types test for useContractRead.ts", () => {
     } as const;
 
     // @ts-expect-error
-    assertType<UseContractReadProps<TAbi, "fn_felt">>(props);
+    assertType<UseReadContractProps<TAbi, "fn_felt">>(props);
   });
 
   it("correct arguments of function for given abi", () => {
@@ -39,7 +39,7 @@ describe("Types test for useContractRead.ts", () => {
       args: "some_bignumberish_value",
     } as const;
 
-    assertType<UseContractReadProps<TAbi, "fn_felt">>(props);
+    assertType<UseReadContractProps<TAbi, "fn_felt">>(props);
   });
 
   it("wrong arguments of function for given abi", () => {
@@ -50,7 +50,7 @@ describe("Types test for useContractRead.ts", () => {
     } as const;
 
     // @ts-expect-error
-    assertType<UseContractReadProps<TAbi, "fn_felt">>(props);
+    assertType<UseReadContractProps<TAbi, "fn_felt">>(props);
   });
 
   it("correct data with correct arguments & function for given abi", () => {
@@ -60,7 +60,7 @@ describe("Types test for useContractRead.ts", () => {
       args: "some_bignumberish_value",
     } as const;
 
-    const { data } = useContractRead(props);
+    const { data } = useReadContract(props);
     assertType<FunctionRet<TAbi, "fn_felt">>(data);
   });
 });
