+++ packages/core/src/hooks/useReadContract.ts
@@ -1,8 +1,6 @@
-import { Chain } from "@starknet-react/chains";
-import { useMemo } from "react";
-import { ArgsOrCalldata, BlockNumber, BlockTag, Contract } from "starknet";
+import { ArgsOrCalldata, BlockNumber } from "starknet";
 
-import { UseQueryProps, UseQueryResult, useQuery } from "~/query";
+import { UseQueryProps, UseQueryResult } from "~/query";
 
 import {
   Abi,
@@ -11,12 +9,9 @@ import {
   FunctionRet,
 } from "abi-wan-kanabi/dist/kanabi";
 
-import { useInvalidateOnBlock } from "./useInvalidateOnBlock";
-import { useNetwork } from "./useNetwork";
+import { ContractReadQueryKey, UseCallProps, useCall } from "./useCall";
 
-const DEFAULT_FETCH_INTERVAL = 5_000;
-
-type ContractReadArgs = {
+type ReadContractArgs = {
   /** The contract's function name. */
   functionName: string;
   /** Read arguments. */
@@ -30,16 +25,16 @@ type Result<
   TFunctionName extends ExtractAbiFunctionNames<TAbi>
 > = FunctionRet<TAbi, TFunctionName>;
 
-/** Options for `useContractRead`. */
-export type UseContractReadProps<
+/** Options for `useReadContract`. */
+export type UseReadContractProps<
   TAbi extends Abi,
   TFunctionName extends ExtractAbiFunctionNames<TAbi>
-> = Omit<ContractReadArgs, "functionName" | "args"> &
+> = Omit<ReadContractArgs, "functionName" | "args"> &
   UseQueryProps<
     Result<TAbi, TFunctionName>,
     Error,
     Result<TAbi, TFunctionName>,
-    ReturnType<typeof queryKey>
+    ReturnType<ContractReadQueryKey>
   > & {
     /** The target contract's ABI.
      *
@@ -75,8 +70,8 @@ export type UseContractReadProps<
     args?: FunctionArgs<TAbi, TFunctionName>;
   };
 
-/** Value returned from `useContractRead`. */
-export type UseContractReadResult<
+/** Value returned from `useReadContract`. */
+export type UseReadContractResult<
   TAbi extends Abi,
   TFunctionName extends ExtractAbiFunctionNames<TAbi>
 > = UseQueryResult<Result<TAbi, TFunctionName>, Error>;
@@ -91,103 +86,12 @@ export type UseContractReadResult<
  *
  * - You must pass `abi` as `const` for autocomplete to work.
  */
-export function useContractRead<
+export function useReadContract<
   TAbi extends Abi,
   TFunctionName extends ExtractAbiFunctionNames<TAbi>
->({
-  abi,
-  address,
-  functionName,
-  args,
-  blockIdentifier = BlockTag.latest,
-  refetchInterval: refetchInterval_,
-  watch = false,
-  enabled: enabled_ = true,
-  ...props
-}: UseContractReadProps<TAbi, TFunctionName>): UseContractReadResult<
-  TAbi,
-  TFunctionName
-> {
-  const { chain } = useNetwork();
-  const contract =
-    abi && address ? new Contract(abi, address).typedv2(abi) : undefined;
-
-  const queryKey_ = useMemo(
-    () =>
-      queryKey({
-        chain,
-        contract,
-        functionName,
-        args: args as ArgsOrCalldata,
-        blockIdentifier,
-      }),
-    [chain, contract, functionName, args, blockIdentifier]
-  );
-
-  const enabled = useMemo(
-    () => Boolean(enabled_ && contract && functionName && args),
-    [enabled_, contract, functionName, args]
-  );
-
-  const refetchInterval =
-    refetchInterval_ ??
-    (blockIdentifier === BlockTag.pending && watch
-      ? DEFAULT_FETCH_INTERVAL
-      : undefined);
-
-  useInvalidateOnBlock({
-    enabled: Boolean(enabled && watch),
-    queryKey: queryKey_,
-  });
-
-  return useQuery({
-    queryKey: queryKey_,
-    queryFn: queryFn({
-      contract,
-      functionName,
-      args: args as ArgsOrCalldata,
-      blockIdentifier,
-    }) as () => Promise<Result<TAbi, TFunctionName>>,
-    refetchInterval,
-    ...props,
-  });
-}
-
-function queryKey({
-  chain,
-  contract,
-  functionName,
-  args,
-  blockIdentifier,
-}: { chain?: Chain; contract?: Contract } & ContractReadArgs) {
-  return [
-    {
-      entity: "readContract",
-      chainId: chain?.name,
-      contract: contract?.address,
-      functionName,
-      args,
-      blockIdentifier,
-    },
-  ] as const;
-}
-
-function queryFn({
-  contract,
-  functionName,
-  args,
-  blockIdentifier,
-}: { contract?: Contract } & ContractReadArgs) {
-  return async () => {
-    if (!contract) throw new Error("contract is required");
-    if (contract.functions[functionName] === undefined) {
-      throw new Error(`function ${functionName} not found in contract`);
-    }
-
-    return contract.call(functionName, args, {
-      parseRequest: true,
-      parseResponse: true,
-      blockIdentifier,
-    });
-  };
+>(props: UseReadContractProps<TAbi, TFunctionName>) {
+  return useCall(props as UseCallProps) as UseReadContractResult<
+    TAbi,
+    TFunctionName
+  >;
 }
